add geofence_radius and custom_image col suport

+++ app/src/main/java/org/actionpath/issues/IssuesDbHelper.java
@@ -10,10 +10,10 @@ import android.util.Log;
  */
 public class IssuesDbHelper extends SQLiteOpenHelper {
 
-    private static String LOG_TAG = IssuesDbHelper.class.getName();
+    private static String TAG = IssuesDbHelper.class.getName();
 
     private static final String DATABASE_NAME = "issues.db";
-    private static final int DATABASE_VERSION = 1;
+    private static final int DATABASE_VERSION = 2;
 
     // DB Table consts
     public static final String ISSUES_TABLE_NAME = "issues";
@@ -30,6 +30,7 @@ public class IssuesDbHelper extends SQLiteOpenHelper {
     public static final String ISSUES_PLACE_ID_COL = "place_id";
     public static final String ISSUES_CREATED_AT_COL = "created_at";
     public static final String ISSUES_UPDATED_AT_COL = "updated_at";
+    public static final String ISSUES_GEOFENCE_RADIUS_COL = "geofence_radius";
 
     public static String[] ISSUES_COLUMN_NAMES;
 
@@ -38,7 +39,7 @@ public class IssuesDbHelper extends SQLiteOpenHelper {
                 {ISSUES_ID_COL,ISSUES_STATUS_COL,ISSUES_SUMMARY_COL,ISSUES_DESCRIPTION_COL,
                         ISSUES_ADDRESS_COL,ISSUES_LATITUDE_COL,ISSUES_LONGITUDE_COL,ISSUES_IMAGE_URL_COL,
                         ISSUES_FOLLOWED_COL,ISSUES_GEOFENCE_CREATED_COL,ISSUES_PLACE_ID_COL,
-                        ISSUES_CREATED_AT_COL,ISSUES_UPDATED_AT_COL };
+                        ISSUES_CREATED_AT_COL,ISSUES_UPDATED_AT_COL, ISSUES_GEOFENCE_RADIUS_COL };
     }
 
     // Database creation sql statement
@@ -57,11 +58,12 @@ public class IssuesDbHelper extends SQLiteOpenHelper {
             ISSUES_PLACE_ID_COL + " int, " +
             ISSUES_CREATED_AT_COL + " int, " +
             ISSUES_UPDATED_AT_COL + " int " +
+            ISSUES_GEOFENCE_RADIUS_COL + " int " +
             ");";
 
     public IssuesDbHelper(Context context) {
         super(context, DATABASE_NAME, null, DATABASE_VERSION);
-        Log.i(LOG_TAG, "Creating new IssuesDbHelper for " + context.getPackageName());
+        Log.i(TAG, "Creating new IssuesDbHelper for " + context.getPackageName());
     }
 
     @Override
@@ -71,11 +73,15 @@ public class IssuesDbHelper extends SQLiteOpenHelper {
 
     @Override
     public void onUpgrade(SQLiteDatabase db, int oldVersion, int newVersion) {
-        Log.w(LOG_TAG,
-                "Upgrading database from version " + oldVersion + " to "
-                        + newVersion + ", which will destroy all old data");
-        db.execSQL("DROP TABLE IF EXISTS " + ISSUES_TABLE_NAME);
-        onCreate(db);
+        Log.w(TAG, "Upgrading database from version " + oldVersion + " to " + newVersion);
+        if(oldVersion==1 && newVersion>=2){
+            db.execSQL("ALTER TABLE "+ISSUES_TABLE_NAME+" ADD COLUMN "+ISSUES_GEOFENCE_RADIUS_COL+" int default 500;");
+            Log.i(TAG,"Upgraded "+ISSUES_TABLE_NAME+" from v1 to v2");
+        } else {
+            db.execSQL("DROP TABLE IF EXISTS " + ISSUES_TABLE_NAME);
+            onCreate(db);
+            Log.i(TAG, "Created " + ISSUES_TABLE_NAME);
+        }
     }
 
 }
